move room sorting logic to ParticipantsList
